add criteria and relate criteria to sub criteria

+++ database/migrations/2025_04_11_094736_create_criterias_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('code')->unique();
             $table->string('name');
-            $table->integer('weight');
+            $table->decimal('weight');
             $table->enum('type', ['Benefit', 'Cost']);
             $table->timestamps();
         });
