+++ database/migrations/2025_04_11_101441_create_assesments_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('employee_id');
             $table->unsignedBigInteger('user_id');
             $table->string('period');
-            $table->integer('score');
+            $table->decimal('score');
             $table->timestamps();
 
             $table->foreign('employee_id')->references('id')->on('employees')->onUpdate('cascade')->onDelete('cascade');
